improve

+++ ptunwrap.c
@@ -9,6 +9,7 @@
 #include <sys/wait.h>
 #include <termios.h>
 #include <unistd.h>
+#include <stdlib.h>
 
 #define FAIL_WITH_PPID
 #include "fail.h"
@@ -20,13 +21,18 @@ static int siglist[] = {
     SIGINT,
     SIGQUIT,
     SIGTERM,
+    SIGUSR1,
+    SIGUSR2,
 };
 
 static volatile int uppid = 0;
 static volatile int downpid = 0;
+static struct termios oldterm;
+
 static void close_handler(int sig);
-static int start_upstream(int up, int down);
-static int start_downstream(int up, int down);
+static int start_upstream(int up);
+static int start_downstream(int down);
+static void restore_term();
 
 int main(int argc, char *argv[]) {
     if(argc < 3) fail_err(EINVAL);
@@ -34,12 +40,7 @@ int main(int argc, char *argv[]) {
     int up = ensure(open(argv[1], O_WRONLY));
     int down = ensure(open(argv[2], O_RDONLY));
 
-    struct sigaction close_act;
-    memset(&close_act, 0, sizeof(close_act));
-    close_act.sa_handler = close_handler;
-
-    struct termios oldterm, newterm;
-
+    struct termios newterm;
     if(isatty(STDIN_FILENO)) {
         ensure(tcgetattr(STDIN_FILENO, &oldterm));
         newterm = oldterm;
@@ -47,11 +48,20 @@ int main(int argc, char *argv[]) {
         ensure(tcsetattr(STDIN_FILENO, TCSADRAIN, &newterm));
     }
 
-    uppid = start_upstream(up, down);
-    downpid = start_downstream(up, down);
+    uppid = start_upstream(up);
+    downpid = start_downstream(down);
     ensure(close(up));
     ensure(close(down));
 
+    if(atexit(restore_term) != 0) {
+        kill(uppid, SIGKILL);
+        kill(downpid, SIGKILL);
+        fail_err(EBADSLT);
+    }
+
+    struct sigaction close_act;
+    memset(&close_act, 0, sizeof(close_act));
+    close_act.sa_handler = close_handler;
     for (int i = 0; i < sizeof(siglist) / sizeof(siglist[0]); ++i) {
         ensure(sigaction(siglist[i], &close_act, NULL));
     }
@@ -60,27 +70,30 @@ int main(int argc, char *argv[]) {
         int rc = waitpid(-1, NULL, 0);
         IF_err(rc) {
             if(errno == ECHILD) break;
-        } else if(rc == uppid || rc == downpid) {
-            kill(uppid, SIGKILL);
-            kill(downpid, SIGKILL);
+        } else if(rc == uppid) {
+            uppid = 0;
+            if(downpid != 0) kill(downpid, SIGKILL);
+        } else if(rc == downpid) {
+            downpid = 0;
+            if(uppid != 0) kill(uppid, SIGKILL);
         }
     }
-
-    if(isatty(STDIN_FILENO)) ensure(tcsetattr(STDIN_FILENO, TCSADRAIN, &oldterm));
 
     return 0;
 }
 
 static void close_handler(int sig) {
-    kill(uppid, SIGKILL);
-    kill(downpid, SIGKILL);
+    if(uppid != 0) kill(uppid, SIGKILL);
+    if(downpid != 0) kill(downpid, SIGKILL);
 }
 
-static int start_upstream(int up, int down) {
+static void restore_term() {
+    if(isatty(STDIN_FILENO)) ensure(tcsetattr(STDIN_FILENO, TCSADRAIN, &oldterm));
+}
+
+static int start_upstream(int up) {
     int ret = ensure(fork());
     if(ret == 0) {
-        ensure(close(STDOUT_FILENO));
-        ensure(close(down));;
         char buf[LOCAL_BUF_SIZE];
         while(1) {
             char *bufptr = buf;
@@ -96,11 +109,9 @@ static int start_upstream(int up, int down) {
     return ret;
 }
 
-static int start_downstream(int up, int down) {
+static int start_downstream(int down) {
     int ret = ensure(fork());
     if(ret == 0) {
-        ensure(close(STDIN_FILENO));
-        ensure(close(up));
         char buf[LOCAL_BUF_SIZE];
         while(1) {
             char *bufptr = buf;
